Add function `from_file` to DNBinaryFormat class
This function takes a `str` or `pathlib.Path` object and converts it
into a `pathlib.Path` object if it isn't one already. A new instance of
the DNBinaryFormat class is then created by reading the `Path` bytes
into `DeserializingStream`.

This function is intended to facilitate the creation of a DNBinaryFormat
instance.